no Jira - correct NPE in test case - checking for existence of test output file - if it's not there, then NPE.  Discovered on Linux.  Doesn't happen on Windows - probably because the file is always there - there's another test case that creates it, and the tests could run in a different order between Windows and Linux.
git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk@819716 13f79535-47bb-0310-9956-ffa450edef68

+++ uimaj-core/src/test/java/org/apache/uima/analysis_engine/impl/AnalysisEngine_implTest.java
@@ -350,7 +350,7 @@ public class AnalysisEngine_implTest extends TestCase {
 
       // test aggregate TAE containing a CAS Consumer
       File outFile = JUnitExtension.getFile("CpmOutput.txt");
-      if(outFile.exists()) {
+      if(outFile != null && outFile.exists()) {
         //outFile.delete() //can't be relied upon.  Instead set file to zero length.
         FileOutputStream fos = new FileOutputStream(outFile, false);
         fos.close();
